// src/lib/backend/api/api.ts
export interface V1AdminProductsGet200Response {
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200Response
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200Response
     */
    'message': string | null;
    /**
     * 
     * @type {V1AdminProductsGet200ResponseData}
     * @memberof V1AdminProductsGet200Response
     */
    'data': V1AdminProductsGet200ResponseData;
}
/**
 * 
 * @export
 * @interface V1AdminProductsGet200ResponseData
 */
export interface V1AdminProductsGet200ResponseData {
    /**
     * 
     * @type {Array<V1AdminProductsGet200ResponseDataDataInner>}
     * @memberof V1AdminProductsGet200ResponseData
     */
    'data': Array<V1AdminProductsGet200ResponseDataDataInner>;
    /**
     * 
     * @type {number}
     * @memberof V1AdminProductsGet200ResponseData
     */
    'count': number;
}
/**
 * 
 * @export
 * @interface V1AdminProductsGet200ResponseDataDataInner
 */
export interface V1AdminProductsGet200ResponseDataDataInner {
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'deletedAt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'expiresAt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'type'?: V1AdminProductsGet200ResponseDataDataInnerTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'category'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'summary'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'currency'?: V1AdminProductsGet200ResponseDataDataInnerCurrencyEnum;
    /**
     * 
     * @type {number}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'price'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'buyMinimumQuantity'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'buyMaximumQuantity'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'buyGroupOfQuantity'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'sku'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'stockQuantity'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsGet200ResponseDataDataInner
     */
    'visiblility'?: V1AdminProductsGet200ResponseDataDataInnerVisiblilityEnum;
}